fix(ci): split redact snapshot restore coverage
(cherry picked from commit aeb2adf240)
